296. Validações parte 1 - Controle de fluxos

+++ app/Http/Controllers/MarcaController.php
@@ -55,6 +55,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -84,7 +87,13 @@ class MarcaController extends Controller
         print_r($marca->getAttributes()); //dados antigos
         */
         //$marca->update($request->all());
+
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
+
         $marca->update($request->all());
 
         return $marca;
@@ -99,6 +108,11 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
+
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso'];
         //return response()->noContent();
